leetcode-java/src/main/java/org/year_2024/may/medium/AddTwoNumbers.java
package org.year_2024.may.medium;

import org.year_2024.may.helpers.ListNode;

public class AddTwoNumbers {
    public static ListNode addTwoNumbers(ListNode l1, ListNode l2) {
        ListNode dummyHead = new ListNode(0);
        ListNode p = l1;
        ListNode q = l2;
        ListNode curr = dummyHead;

        int carry = 0;

        while (p != null || q != null) {
            int sum = carry + getAnInt(p) + getAnInt(q);

            carry = sum / 10;
            curr.next = new ListNode(sum % 10);
            curr = curr.next;

            if (p != null) {
                p = p.next;
            }
            if (q != null) {
                q = q.next;
            }
        }

        if (carry > 0) {
            curr.next = new ListNode(carry);
        }
        return dummyHead.next;
    }

    private static int getAnInt(ListNode p) {
        return p != null ? p.val : 0;
    }
}
